Make TTL of file content table entries settable from command line argument
This change also increases the upper bound of TTL to `ushort`, but the default TTL is still set to `byte.MaxValue`.

Office has been experiencing FCT trashing because the TTL is too small as they are now sharing FCT over multiple enlistments. Also, consider that Office developers do obuild a lot of times, and so when they do ohome afterward, they are losing a lot of entries in FCT.

+++ Public/Src/Utilities/Configuration/ICacheConfiguration.cs
@@ -93,7 +93,7 @@ namespace BuildXL.Utilities.Configuration
         /// <summary>
         /// The time to to live (number of build iterations) for entries to be retained without use before eviction
         /// </summary>
-        byte? FileContentTableEntryTimeToLive { get; set; }
+        ushort? FileContentTableEntryTimeToLive { get; set; }
 
         /// <summary>
         /// Allows for fetching cached graph from content cache.
